refactor project/samples_test into separate files
